Fix parser to handle edge cases

+++ philo_new/inc/philo.h
@@ -6,7 +6,7 @@
 /*   By: ogonzale <ogonzale@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 15:58:55 by ogonzale          #+#    #+#             */
-/*   Updated: 2022/11/02 16:51:53 by ogonzale         ###   ########.fr       */
+/*   Updated: 2022/11/03 16:40:01 by ogonzale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,30 +19,30 @@ typedef struct s_data	t_data;
 
 typedef struct s_philo
 {
-	int			philo_num;
-	int			right_fork;
-	int			left_fork;
-	long int	last_meal;
-	int			needs_to_eat;
-	long int	time_of_start;
-	t_data		*data;
-}				t_philo;
+	int				philo_num;
+	int				right_fork;
+	int				left_fork;
+	long long int	last_meal;
+	int				needs_to_eat;
+	long long int	time_of_start;
+	t_data			*data;
+}					t_philo;
 
 typedef struct s_data
 {
-	int				num_philos;
-	long int		time_to_die;
-	long int		time_to_eat;
-	long int		time_to_sleep;
-	long int		max_meals;
-	int				death;
-	long int		time_absolute;
-	long int		time_relative;
-	t_philo			*philo;
-	pthread_mutex_t	m_death;
-	pthread_mutex_t	*m_forks;
-	pthread_mutex_t	m_print;
-}					t_data;
+	int					num_philos;
+	long long int		time_to_die;
+	long long int		time_to_eat;
+	long long int		time_to_sleep;
+	long long int		max_meals;
+	int					death;
+	long long int		time_absolute;
+	long long int		time_relative;
+	t_philo				*philo;
+	pthread_mutex_t		m_death;
+	pthread_mutex_t		*m_forks;
+	pthread_mutex_t		m_print;
+}						t_data;
 
 /* parse.c */
 
@@ -50,6 +50,6 @@ int	ft_parse(int argc, char *argv[], t_data *data);
 
 /* ft_atoi_mod.c */
 
-int	ft_atoi_mod(const char *nptr);
+long long int	ft_atoi_mod(const char *nptr);
 
 #endif
